+++ desktop/libs/Constants.pas
@@ -15,6 +15,7 @@ const DEFAULT_LABEL_VERSION = 'Version';
 const DEFAULT_ERROR_TITLE = 'Error';
 const LANGUAGE_LIST_FILE = 'available_languages';
 const SEPARATOR_MENU = '-';
+const PARAM_VERSION_NAME = 'VERSION';
 
 implementation